Refactor component structure and file paths
