--- src/components/commons/Column.jsx
@@ -1,6 +1,0 @@
-export function Columns({ children }) {
-  return <div className="columns is-fullheight">{children}</div>;
-}
-export function Column({ children, ratio }) {
-  return <div className={`column is-${ratio}`}>{children}</div>;
-}
